내정보가져오기 -회원정보

+++ src/main/java/com/example/cucucook/service/MemberService.java
@@ -18,7 +18,7 @@ public interface MemberService {
 
     boolean checkEmailExists(String email);
 
-    Member getMember(Long memberId);
+    Member getMember(int memberId);
 
     void updateMember(Member member);
